requirements.txt
openai
streamlit
pymupdf
python-docx
pandas
numpy
requests
plotly
